+++ src/app/code/community/Shopgate/Cloudapi/Model/Api2/Customers/Rest.php
@@ -22,4 +22,59 @@
 
 class Shopgate_Cloudapi_Model_Api2_Customers_Rest extends Shopgate_Cloudapi_Model_Api2_Resource
 {
+    /**
+     * Get customer resource validator instance
+     *
+     * @return Shopgate_Cloudapi_Model_Api2_Validator
+     */
+    protected function _getValidator()
+    {
+        return Mage::getModel('shopgate_cloudapi/api2_validator', array('resource' => $this));
+    }
+
+    /**
+     * Sets detailed validation errors to be returned by the address endpoints
+     *
+     * @param Shopgate_Cloudapi_Model_Api2_Validator $validator
+     *
+     * @return Shopgate_Cloudapi_Model_Api2_Customers_Rest
+     */
+    protected function setDetailedErrors($validator)
+    {
+        foreach ($validator->getDetailedErrors() as $code => $errors) {
+            $this->_errorMessage(
+                '',
+                Mage_Api2_Model_Server::HTTP_BAD_REQUEST,
+                array('path' => $code, 'messages' => $errors)
+            );
+        }
+
+        return $this;
+    }
+
+    /**
+     * Bypasses the exception state and passes down invalidation errors
+     *
+     * @throws Zend_Controller_Response_Exception
+     * @throws Exception
+     */
+    public function sendInvalidationResponse()
+    {
+        $this->getResponse()->setHttpResponseCode(Mage_Api2_Model_Server::HTTP_BAD_REQUEST);
+
+        return array('messages' => $this->getResponse()->getMessages());
+    }
+
+    /**
+     * @inheritdoc
+     * @return Shopgate_Cloudapi_Model_Acl_Customers_Filter
+     */
+    public function getFilter()
+    {
+        if (!$this->filter) {
+            $this->filter = Mage::getModel('shopgate_cloudapi/acl_customers_filter', $this);
+        }
+
+        return $this->filter;
+    }
 }