+++ CHANGELOG.md
@@ -11,10 +11,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Endpoint for customer address creation
 - Endpoint for customer address deletion
 - Endpoint for customer address update
+- Endpoint for customer data update
 - Shell script to update our REST attributes / roles
 - Support of app-only coupons (except when EE Customer Segments is installed)
 ### Changed
 - Initial installed admin role, it no longer uses all resources
+- Endpoint for basic customer data now retrieves the user group, but not the addresses
 
 ## [3.1.3] - 2018-05-30
 ### Fixed